Removed the early poc configuration file options
- Also removed `--agent-mode` as it is the only mode now.
